Fix clippy, move a couple more methods
